fix compiler warnings for 2.13.2
Most of the new warnings were for widening conversions from
long to double. Since cross-building has been disabled, all
warnings have been cleared up with no issues. Enabled fatal
warnings to help keep the build clean.

+++ atlas-chart/src/main/scala/com/netflix/atlas/chart/graphics/Scales.scala
@@ -95,8 +95,8 @@ object Scales {
   }
 
   /** Factory for creating a mapping for time values. */
-  def time(d1: Double, d2: Double, step: Long, r1: Int, r2: Int): LongScale = {
-    val dr = (d2 - d1) / step
+  def time(d1: Long, d2: Long, step: Long, r1: Int, r2: Int): LongScale = {
+    val dr = (d2 - d1).toDouble / step
     val pixelsPerStep = (r2 - r1) / dr
     v => ((v - d1) / step * pixelsPerStep).toInt + r1
   }
